FESTOU-116: Desenvolver lógica para cadastro de espaço com várias telas
