+++ lib/src/core/providers/application_providers.dart
@@ -1,7 +1,6 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:git_flutter_festou/src/core/ui/festou_nav_global_key.dart';
-import 'package:git_flutter_festou/src/features/login/login_vm.dart';
 import 'package:git_flutter_festou/src/repositories/feedback/feedback_firestore_repository.dart';
 import 'package:git_flutter_festou/src/repositories/feedback/feedback_firestore_repository_impl.dart';
 import 'package:git_flutter_festou/src/repositories/images/images_storage_repository.dart';
