+++ lib/src/features/bottomNavBar/home/home_page.dart
@@ -8,7 +8,6 @@ import 'package:git_flutter_festou/src/features/register/space/space_register_pa
 import 'package:git_flutter_festou/src/features/show%20spaces/all%20space%20mvvm/all_spaces_state.dart';
 import 'package:git_flutter_festou/src/features/show%20spaces/all%20space%20mvvm/all_spaces_vm.dart';
 import 'package:git_flutter_festou/src/features/bottomNavBar/home/widgets/my_last_seen_spaces.dart';
-import 'package:git_flutter_festou/src/features/show%20spaces/my%20space%20mvvm/my_spaces_page.dart';
 
 class HomePage extends ConsumerStatefulWidget {
   final String? previousRoute;
