+++ lib/src/features/bottomNavBar/profile/profile_vm.dart
@@ -1,5 +1,3 @@
-import 'package:flutter/material.dart';
-import 'package:git_flutter_festou/src/core/exceptions/auth_exception.dart';
 import 'package:git_flutter_festou/src/core/exceptions/repository_exception.dart';
 import 'package:git_flutter_festou/src/core/fp/either.dart';
 import 'package:git_flutter_festou/src/core/providers/application_providers.dart';
