+++ lib/src/features/register/reserva/reserva_register_state.dart
@@ -1,8 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/material.dart';
-import 'dart:io';
-import 'package:git_flutter_festou/src/models/space_model.dart';
-import 'package:git_flutter_festou/src/models/space_with_image_model.dart';
 
 enum ReservationRegisterStateStatus { initial, success, error }
