+++ lib/src/features/register/reserva/reserva_temporary_vm.dart
@@ -1,12 +1,9 @@
-import 'dart:developer';
 
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:git_flutter_festou/src/core/exceptions/repository_exception.dart';
 import 'package:git_flutter_festou/src/core/fp/either.dart';
-import 'package:git_flutter_festou/src/core/providers/application_providers.dart';
 import 'package:git_flutter_festou/src/features/register/reserva/reserva_register_state.dart';
 import 'package:git_flutter_festou/src/features/register/reserva/reserva_register_vm.dart';
-import 'package:git_flutter_festou/src/features/register/reserva/reserva_temporary_state.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:uuid/uuid.dart';
