+++ lib/src/features/register/space/space temporary/pages/revisao.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:git_flutter_festou/src/core/ui/helpers/messages.dart';
+import 'package:git_flutter_festou/src/features/register/space/space%20temporary/pages/new_space_register_vm.dart';
 import 'package:git_flutter_festou/src/features/register/space/space_register_state.dart';
 import 'package:git_flutter_festou/src/features/register/space/space_register_vm.dart';
-import 'package:git_flutter_festou/src/features/register/space/widgets/pages/new_space_register_state.dart';
-import 'package:git_flutter_festou/src/features/register/space/widgets/pages/new_space_register_vm.dart';
 
 class Revisao extends ConsumerStatefulWidget {
   const Revisao({super.key});
