--- lib/src/features/register/space/widgets/pages/adicione_fotos.dart
@@ -1,119 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:git_flutter_festou/src/features/register/space/space_register_vm.dart';
-import 'package:git_flutter_festou/src/features/register/space/widgets/pages/new_space_register_vm.dart';
-import 'package:git_flutter_festou/src/features/register/space/widgets/pages/pronto_quetal.dart';
-
-class AdicioneFotos extends ConsumerStatefulWidget {
-  const AdicioneFotos({super.key});
-
-  @override
-  ConsumerState<AdicioneFotos> createState() => _AdicioneFotosState();
-}
-
-class _AdicioneFotosState extends ConsumerState<AdicioneFotos> {
-  @override
-  Widget build(BuildContext context) {
-    final spaceRegister = ref.watch(spaceRegisterVmProvider.notifier);
-    return Scaffold(
-      body: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Padding(
-            padding: const EdgeInsets.all(25.0),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                InkWell(
-                  onTap: () {},
-                  child: Container(
-                    padding: const EdgeInsets.all(7.0),
-                    decoration: BoxDecoration(
-                        color: Colors.white,
-                        border: Border.all(),
-                        borderRadius: BorderRadius.circular(20)),
-                    child: const Text(
-                      'Salvar e sair',
-                    ),
-                  ),
-                ),
-                InkWell(
-                  onTap: () {},
-                  child: Container(
-                    padding: const EdgeInsets.all(7.0),
-                    decoration: BoxDecoration(
-                        color: Colors.white,
-                        border: Border.all(),
-                        borderRadius: BorderRadius.circular(20)),
-                    child: const Text(
-                      'Dúvidas?',
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-          const Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                'Adicione algumas fotos da sua acomodaçao((tipo de espaço))',
-                style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20),
-              ),
-              Text(
-                  'voce precisara de cinco fotos para começar. voce pode adicionar outras imagens ou faer alterações mais tarde.'),
-            ],
-          ),
-          const SizedBox(
-            height: 20,
-          ),
-          ElevatedButton(
-              onPressed: () => spaceRegister.pickImage(),
-              child: const Text('Adicionar fotos')),
-          ElevatedButton(
-              onPressed: () {}, child: const Text('Tirar novas fotos')),
-          const SizedBox(
-            height: 20,
-          ),
-          Padding(
-            padding: const EdgeInsets.all(15.0),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                InkWell(
-                  onTap: () => Navigator.pop(context),
-                  child: const Text(
-                    'Voltar',
-                    style: TextStyle(
-                      decoration: TextDecoration.underline,
-                    ),
-                  ),
-                ),
-                InkWell(
-                  onTap: () => Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) => const ProntoQuetal(),
-                    ),
-                  ),
-                  child: Container(
-                    decoration: BoxDecoration(
-                        color: Colors.black,
-                        border: Border.all(),
-                        borderRadius: BorderRadius.circular(10)),
-                    padding: const EdgeInsets.symmetric(
-                        horizontal: 15, vertical: 10),
-                    child: const Text(
-                      'Avançar',
-                      style: TextStyle(color: Colors.white),
-                    ),
-                  ),
-                )
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
