--- lib/src/features/register/space/widgets/pages/tipo_espaco.dart
@@ -1,148 +1,0 @@
-import 'dart:developer';
-
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:git_flutter_festou/src/features/register/space/widgets/pages/localizacao.dart';
-import 'package:git_flutter_festou/src/features/register/space/widgets/pages/new_space_register_vm.dart';
-import 'package:git_flutter_festou/src/features/register/space/widgets/type_panel.dart';
-
-class TipoEspaco extends ConsumerStatefulWidget {
-  const TipoEspaco({super.key});
-
-  @override
-  ConsumerState<TipoEspaco> createState() => _TipoEspacoState();
-}
-
-class _TipoEspacoState extends ConsumerState<TipoEspaco> {
-  @override
-  Widget build(BuildContext context) {
-    final newSpaceRegister = ref.read(newSpaceRegisterVmProvider.notifier);
-
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text(''),
-      ),
-      body: Column(
-        children: [
-          Padding(
-            padding: const EdgeInsets.all(25.0),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                InkWell(
-                  onTap: () {},
-                  child: Container(
-                    padding: const EdgeInsets.all(7.0),
-                    decoration: BoxDecoration(
-                        color: Colors.white,
-                        border: Border.all(),
-                        borderRadius: BorderRadius.circular(20)),
-                    child: const Text(
-                      'Salvar e sair',
-                    ),
-                  ),
-                ),
-                InkWell(
-                  onTap: () {},
-                  child: Container(
-                    padding: const EdgeInsets.all(7.0),
-                    decoration: BoxDecoration(
-                        color: Colors.white,
-                        border: Border.all(),
-                        borderRadius: BorderRadius.circular(20)),
-                    child: const Text(
-                      'Dúvidas?',
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-          Column(
-            children: [
-              const Text(
-                'Qual das seguintes opções descreve melhor seu espaço?',
-                style: TextStyle(fontWeight: FontWeight.bold, fontSize: 18),
-              ),
-              TypePanel(
-                text: 'Selecione o TIPO de espaço',
-                onTypePressed: (value) {
-                  log('onTypePressed: $value');
-                  newSpaceRegister.addOrRemoveType(value);
-                },
-                selectedTypes: const [],
-              ),
-            ],
-          ),
-          Padding(
-            padding: const EdgeInsets.all(15.0),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                InkWell(
-                  onTap: () => Navigator.pop(context),
-                  child: const Text(
-                    'Voltar',
-                    style: TextStyle(
-                      decoration: TextDecoration.underline,
-                    ),
-                  ),
-                ),
-                InkWell(
-                  onTap: () => Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) => const Localizacao(),
-                    ),
-                  ),
-                  child: Container(
-                    decoration: BoxDecoration(
-                        color: Colors.black,
-                        border: Border.all(),
-                        borderRadius: BorderRadius.circular(10)),
-                    padding: const EdgeInsets.symmetric(
-                        horizontal: 15, vertical: 10),
-                    child: const Text(
-                      'Avançar',
-                      style: TextStyle(color: Colors.white),
-                    ),
-                  ),
-                )
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class GridItem extends StatelessWidget {
-  final IconData icon;
-  final String text;
-
-  const GridItem({Key? key, required this.icon, required this.text})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      elevation: 5.0,
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Icon(
-            icon,
-            size: 40.0,
-            color: Colors.blue,
-          ),
-          const SizedBox(height: 8.0),
-          Text(
-            text,
-            style: const TextStyle(fontSize: 16.0),
-          ),
-        ],
-      ),
-    );
-  }
-}
